test(meshtastic): consolidate utility test files into metadata suite
Merge health, node name refresh, and probe coverage tests into `test_meshtastic_utils_metadata.py` to reduce file fragmentation.

Changes:
- Deleted `test_meshtastic_utils_health.py`, `test_meshtastic_utils_node_name_refresh.py`, and `test_meshtastic_utils_probe_coverage.py`.
- Migrated all test cases and helper classes to `test_meshtastic_utils_metadata.py`.
- Updated `reset_meshtastic_relay_state` fixture to include comprehensive cleanup of global state variables used by the absorbed tests.
