+++ tests/test_meshtastic_utils_metadata.py
@@ -12,17 +12,28 @@ Tests the Meshtastic client functionality including:
 
 import asyncio
 import contextlib
+import logging
 import threading
 import unittest
+from concurrent.futures import Future
 from concurrent.futures import TimeoutError as ConcurrentTimeoutError
 from types import SimpleNamespace
-from typing import Any
-from unittest.mock import MagicMock, Mock, patch
+from typing import Any, cast
+from unittest.mock import ANY, AsyncMock, MagicMock, Mock, patch
 
 import pytest
 
+import mmrelay.meshtastic_utils as mu
+from mmrelay.constants.config import (
+    DEFAULT_HEALTH_CHECK_ENABLED,
+    DEFAULT_NODEDB_REFRESH_INTERVAL,
+)
 from mmrelay.constants.network import (
+    CONFIG_KEY_CONNECTION_TYPE,
+    CONNECTION_TYPE_BLE,
     CONNECTION_TYPE_TCP,
+    DEFAULT_MESHTASTIC_OPERATION_TIMEOUT,
+    INITIAL_HEALTH_CHECK_DELAY,
     METADATA_WATCHDOG_SECS,
 )
 from mmrelay.meshtastic_utils import (
@@ -115,6 +126,61 @@ def reset_meshtastic_relay_state(monkeypatch):
     monkeypatch.setattr(
         "mmrelay.meshtastic_utils._health_probe_request_deadlines",
         {},
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.config",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.meshtastic_client",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.meshtastic_iface",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils._metadata_future",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils._metadata_future_started_at",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils._metadata_executor_degraded",
+        False,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.reconnecting",
+        False,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.shutting_down",
+        False,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.RELAY_START_TIME",
+        0,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils._ble_future",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils._ble_executor",
+        None,
         raising=False,
     )
 
@@ -841,3 +907,833 @@ def test_resolve_plugin_timeout(cfg, default, expected):
     """Test _resolve_plugin_timeout with various configurations."""
     result = _resolve_plugin_timeout(cfg, default)
     assert result == expected
+
+
+# ---------------------------------------------------------------------------
+# Absorbed from tests/test_meshtastic_utils_health.py
+# ---------------------------------------------------------------------------
+
+
+def _make_health_config(
+    connection_type: str = CONNECTION_TYPE_TCP,
+    enabled: bool = True,
+    heartbeat: int = 60,
+    initial_delay: float | None = None,
+    probe_timeout: float | None = None,
+) -> dict[str, Any]:
+    """
+    Builds a nested configuration dictionary for meshtastic connection health checks.
+
+    Parameters:
+        connection_type (str): Connection transport type (e.g., "tcp" or "ble"). Defaults to "tcp".
+        enabled (bool): Whether health checks are enabled. Defaults to True.
+        heartbeat (int): Heartbeat interval in seconds used for health check scheduling. Defaults to 60.
+        initial_delay (float | None): Optional delay before first health check.
+        probe_timeout (float | None): Optional timeout per health probe.
+
+    Returns:
+        dict: Configuration mapping with keys "meshtastic" -> {"connection_type", "health_check": {"enabled", "heartbeat_interval"}}.
+    """
+    health_check = {"enabled": enabled, "heartbeat_interval": heartbeat}
+    if initial_delay is not None:
+        health_check["initial_delay"] = initial_delay
+    if probe_timeout is not None:
+        health_check["probe_timeout"] = probe_timeout
+
+    return {
+        "meshtastic": {
+            CONFIG_KEY_CONNECTION_TYPE: connection_type,
+            "health_check": health_check,
+        }
+    }
+
+
+class SleepAndShutdown:
+    """
+    Helper to trigger shutdown after a specified number of sleep calls.
+
+    This is used to test the health check loop which has an initial delay
+    before the first check, followed by a loop sleep after each check.
+    """
+
+    def __init__(self, shutdown_after: int = 1):
+        """
+        Initialize the helper.
+
+        Parameters:
+            shutdown_after (int): Number of sleep calls before triggering shutdown.
+                Defaults to 1 (shutdown on first sleep, for tests that don't need initial delay).
+        """
+        self.sleep_count = 0
+        self.shutdown_after = shutdown_after
+
+    def __call__(self, _seconds: float) -> None:
+        """
+        Increment sleep counter and trigger shutdown after configured count.
+
+        Parameters:
+            _seconds (float): Ignored --- present only to match the asyncio.sleep signature.
+        """
+        self.sleep_count += 1
+        if self.sleep_count >= self.shutdown_after:
+            mu.shutting_down = True
+        return None
+
+
+def _sleep_and_shutdown(_seconds: float) -> None:
+    """
+    Mark the application as shutting down; intended as a placeholder to use where asyncio.sleep is expected.
+
+    Parameters:
+        _seconds (float): Ignored --- present only to match the asyncio.sleep signature.
+    """
+    mu.shutting_down = True
+    return None
+
+
+@pytest.mark.asyncio
+async def test_check_connection_health_disabled_returns():
+    mu.config = _make_health_config(enabled=False)
+
+    with patch("mmrelay.meshtastic_utils.logger") as mock_logger:
+        await check_connection()
+
+    mock_logger.info.assert_called_with(
+        "Connection health checks are disabled in configuration"
+    )
+
+
+@pytest.mark.asyncio
+async def test_check_connection_ble_skips_health_checks():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_BLE)
+    mu.meshtastic_client = MagicMock()
+
+    with patch("mmrelay.meshtastic_utils.logger") as mock_logger:
+        await check_connection()
+
+    mock_logger.debug.assert_any_call(
+        "BLE connection uses real-time disconnection detection; periodic health checks disabled"
+    )
+
+
+@pytest.mark.asyncio
+async def test_check_connection_metadata_probe_succeeds():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = MagicMock()
+    mu.meshtastic_client.localNode.onAckNak = Mock()
+
+    executor = Mock()
+
+    def _submit(fn, *args, **kwargs):
+        probe_future: Future[None] = Future()
+        fn(*args, **kwargs)
+        probe_future.set_result(None)
+        return probe_future
+
+    executor.submit.side_effect = _submit
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch("mmrelay.meshtastic_utils._get_metadata_executor", return_value=executor),
+        patch("mmrelay.meshtastic_utils._probe_device_connection") as mock_probe,
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    executor.submit.assert_called_once()
+    mock_probe.assert_called_once()
+    assert mock_probe.call_args.args[0] is mu.meshtastic_client
+    assert mock_probe.call_args.args[1] == DEFAULT_MESHTASTIC_OPERATION_TIMEOUT
+    mock_logger.error.assert_not_called()
+
+
+@pytest.mark.asyncio
+async def test_check_connection_uses_configured_initial_delay():
+    mu.config = _make_health_config(
+        connection_type=CONNECTION_TYPE_TCP,
+        heartbeat=5,
+        initial_delay=2.5,
+    )
+    mu.meshtastic_client = None
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with patch(
+        "mmrelay.meshtastic_utils.asyncio.sleep",
+        new_callable=AsyncMock,
+        side_effect=sleep_handler,
+    ) as mock_sleep:
+        await check_connection()
+
+    assert mock_sleep.call_count == 2
+    assert mock_sleep.call_args_list[0].args[0] == 2.5
+    assert mock_sleep.call_args_list[1].args[0] == 5
+
+
+@pytest.mark.asyncio
+async def test_check_connection_uses_configured_probe_timeout():
+    mu.config = _make_health_config(
+        connection_type=CONNECTION_TYPE_TCP, probe_timeout=7.5
+    )
+    mu.meshtastic_client = MagicMock()
+    mu.meshtastic_client.localNode.onAckNak = Mock()
+
+    executor = Mock()
+
+    def _submit(fn, *args, **kwargs):
+        probe_future: Future[None] = Future()
+        fn(*args, **kwargs)
+        probe_future.set_result(None)
+        return probe_future
+
+    async def _wait_for_passthrough(awaitable, timeout):
+        # Intentionally ignore timeout - this is a passthrough for testing
+        _ = timeout
+        return await awaitable
+
+    executor.submit.side_effect = _submit
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch("mmrelay.meshtastic_utils._get_metadata_executor", return_value=executor),
+        patch("mmrelay.meshtastic_utils._probe_device_connection") as mock_probe,
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.wait_for",
+            new_callable=AsyncMock,
+            side_effect=_wait_for_passthrough,
+        ) as mock_wait_for,
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+    ):
+        await check_connection()
+
+    mock_probe.assert_called_once()
+    assert mock_probe.call_args.args[0] is mu.meshtastic_client
+    assert mock_probe.call_args.args[1] == 7.5
+    assert mock_wait_for.call_count == 1
+    assert mock_wait_for.call_args.kwargs["timeout"] == 7.5
+
+
+@pytest.mark.asyncio
+async def test_check_connection_triggers_reconnect_on_probe_failure():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = MagicMock()
+    mu.meshtastic_client.localNode.onAckNak = Mock()
+
+    executor = Mock()
+    probe_future: Future[None] = Future()
+    probe_future.set_exception(Exception("probe failed"))
+    executor.submit.return_value = probe_future
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch("mmrelay.meshtastic_utils._get_metadata_executor", return_value=executor),
+        patch("mmrelay.meshtastic_utils.on_lost_meshtastic_connection") as mock_lost,
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    mock_lost.assert_called_once()
+    mock_logger.error.assert_any_call(
+        "%s connection health check failed: %s",
+        "Tcp",
+        ANY,
+        exc_info=True,
+    )
+
+
+@pytest.mark.asyncio
+async def test_check_connection_tracks_timed_out_probe_until_worker_finishes():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = MagicMock()
+    mu.meshtastic_client.localNode.onAckNak = Mock()
+
+    executor = Mock()
+    probe_future: Future[None] = Future()
+    assert probe_future.set_running_or_notify_cancel()
+    executor.submit.return_value = probe_future
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch("mmrelay.meshtastic_utils._get_metadata_executor", return_value=executor),
+        patch.object(mu, "DEFAULT_MESHTASTIC_OPERATION_TIMEOUT", 0.01),
+        patch("mmrelay.meshtastic_utils.on_lost_meshtastic_connection") as mock_lost,
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.wait_for",
+            new_callable=AsyncMock,
+            side_effect=asyncio.TimeoutError,
+        ),
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    assert mu._metadata_future is probe_future
+    executor.submit.assert_called_once()
+    mock_lost.assert_called_once()
+    mock_logger.error.assert_any_call(
+        "%s connection health check failed: %s",
+        "Tcp",
+        ANY,
+        exc_info=True,
+    )
+
+    probe_future.set_result(None)
+    assert mu._metadata_future is None
+
+
+@pytest.mark.asyncio
+async def test_check_connection_skips_when_metadata_probe_active():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = MagicMock()
+    mu.meshtastic_client.localNode.onAckNak = Mock()
+    mu._metadata_future = Mock()
+    mu._metadata_future.done.return_value = False
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils._get_metadata_executor") as mock_executor,
+        patch("mmrelay.meshtastic_utils.on_lost_meshtastic_connection") as mock_lost,
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    mock_executor.assert_not_called()
+    mock_lost.assert_not_called()
+    mock_logger.debug.assert_any_call(
+        "Skipping connection check - metadata probe already in progress"
+    )
+
+
+@pytest.mark.asyncio
+async def test_check_connection_skips_when_reconnecting():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = MagicMock()
+    mu.reconnecting = True
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    mock_logger.debug.assert_any_call(
+        "Skipping connection check - reconnection in progress"
+    )
+
+
+@pytest.mark.asyncio
+async def test_check_connection_skips_when_no_client():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    mu.meshtastic_client = None
+
+    sleep_handler = SleepAndShutdown(
+        shutdown_after=2
+    )  # Shutdown after initial delay + loop sleep
+    with (
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=sleep_handler,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await check_connection()
+
+    mock_logger.debug.assert_any_call("Skipping connection check - no client available")
+
+
+@pytest.mark.asyncio
+async def test_check_connection_uses_legacy_heartbeat_interval():
+    mu.config = _make_health_config(connection_type=CONNECTION_TYPE_TCP)
+    del mu.config["meshtastic"]["health_check"]["heartbeat_interval"]
+    mu.config["meshtastic"]["heartbeat_interval"] = 5
+    mu.meshtastic_client = None
+
+    # We need to survive the first sleep (initial delay) to reach the loop
+    # where the heartbeat interval is used.
+    sleep_handler = SleepAndShutdown(shutdown_after=2)
+
+    with patch(
+        "mmrelay.meshtastic_utils.asyncio.sleep",
+        new_callable=AsyncMock,
+        side_effect=sleep_handler,
+    ) as mock_sleep:
+        await check_connection()
+
+    # Should be called twice:
+    # 1. Initial delay (INITIAL_HEALTH_CHECK_DELAY)
+    # 2. Heartbeat interval (5)
+    assert mock_sleep.call_count == 2
+
+    # Check the first call specifically (initial delay)
+    assert mock_sleep.call_args_list[0].args[0] == INITIAL_HEALTH_CHECK_DELAY
+
+    # Check the second call specifically (the heartbeat)
+    # call_args_list[1] is the second call, args[0] is the first arg
+    assert mock_sleep.call_args_list[1].args[0] == 5
+
+
+def test_probe_device_connection_handles_admin_response_without_routing():
+    class AckState:
+        def __init__(self):
+            self.receivedAck = False
+            self.receivedNak = False
+            self.receivedImplAck = False
+            self.reset = Mock(side_effect=self._reset)
+
+        def _reset(self):
+            self.receivedAck = False
+            self.receivedNak = False
+            self.receivedImplAck = False
+
+    ack_state = AckState()
+    local_node = SimpleNamespace(nodeNum=12345)
+    local_node.iface = SimpleNamespace(
+        _acknowledgment=ack_state,
+        localNode=local_node,
+    )
+
+    client = SimpleNamespace(
+        localNode=local_node,
+        _acknowledgment=ack_state,
+        waitForAckNak=Mock(),
+    )
+
+    def _send_data_side_effect(*_args, **kwargs):
+        kwargs["onResponse"]({"from": str(local_node.nodeNum), "decoded": {}})
+        return None
+
+    client.sendData = Mock(side_effect=_send_data_side_effect)
+
+    mu._probe_device_connection(client)
+
+    client.sendData.assert_called_once()
+    assert ack_state.reset.call_count >= 1
+    client.waitForAckNak.assert_not_called()
+
+
+def test_probe_device_connection_uses_bounded_ack_timeout():
+    ack_state = SimpleNamespace(
+        receivedAck=False,
+        receivedNak=False,
+        receivedImplAck=False,
+        reset=Mock(),
+    )
+    local_node = SimpleNamespace(nodeNum=12345)
+    local_node.iface = SimpleNamespace(
+        _acknowledgment=ack_state,
+        localNode=local_node,
+    )
+
+    client = SimpleNamespace(
+        localNode=local_node,
+        _acknowledgment=ack_state,
+        sendData=Mock(return_value=None),
+        waitForAckNak=Mock(),
+    )
+
+    with patch("mmrelay.meshtastic_utils.time.sleep", return_value=None):
+        with pytest.raises(TimeoutError):
+            mu._probe_device_connection(client, timeout_secs=0.01)
+
+    client.sendData.assert_called_once()
+    client.waitForAckNak.assert_not_called()
+
+
+def test_probe_device_connection_tracks_request_id_for_health_logging():
+    local_node = SimpleNamespace(nodeNum=12345)
+    client = SimpleNamespace(
+        localNode=local_node,
+        _acknowledgment=None,
+        sendData=Mock(return_value=SimpleNamespace(id=987654321)),
+        waitForAckNak=Mock(),
+    )
+
+    with patch(
+        "mmrelay.meshtastic_utils._track_health_probe_request_id",
+        return_value=987654321,
+    ) as mock_track:
+        mu._probe_device_connection(client, timeout_secs=7.5)
+
+    mock_track.assert_called_once_with(987654321, 7.5)
+    client.waitForAckNak.assert_called_once()
+
+
+# ---------------------------------------------------------------------------
+# Absorbed from tests/test_meshtastic_utils_probe_coverage.py
+# ---------------------------------------------------------------------------
+
+
+class TestGetConnectTimeProbeSettings(unittest.TestCase):
+    def test_ble_returns_disabled(self):
+        enabled, timeout = mu._get_connect_time_probe_settings(
+            {"meshtastic": {}}, CONNECTION_TYPE_BLE
+        )
+        self.assertFalse(enabled)
+        self.assertEqual(timeout, float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT))
+
+    def test_none_config_returns_defaults(self):
+        enabled, timeout = mu._get_connect_time_probe_settings(
+            None, CONNECTION_TYPE_TCP
+        )
+        self.assertEqual(enabled, DEFAULT_HEALTH_CHECK_ENABLED)
+        self.assertEqual(timeout, float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT))
+
+    def test_non_dict_config_returns_defaults(self):
+        enabled, timeout = mu._get_connect_time_probe_settings(
+            "not_a_dict",  # type: ignore[arg-type]
+            CONNECTION_TYPE_TCP,
+        )
+        self.assertEqual(enabled, DEFAULT_HEALTH_CHECK_ENABLED)
+        self.assertEqual(timeout, float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT))
+
+    def test_config_without_meshtastic_key_returns_defaults(self):
+        enabled, timeout = mu._get_connect_time_probe_settings(
+            {"other": {}}, CONNECTION_TYPE_TCP
+        )
+        self.assertEqual(enabled, DEFAULT_HEALTH_CHECK_ENABLED)
+        self.assertEqual(timeout, float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT))
+
+    def test_meshtastic_not_dict_returns_defaults(self):
+        enabled, timeout = mu._get_connect_time_probe_settings(
+            {"meshtastic": "not_a_dict"}, CONNECTION_TYPE_TCP
+        )
+        self.assertEqual(enabled, DEFAULT_HEALTH_CHECK_ENABLED)
+        self.assertEqual(timeout, float(DEFAULT_MESHTASTIC_OPERATION_TIMEOUT))
+
+
+class TestScheduleConnectTimeCalibrationProbe(unittest.TestCase):
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_client_no_local_node_returns_early(self, mock_settings):
+        mock_settings.return_value = (True, 30.0)
+        client = Mock(spec=["sendData"])
+        client.sendData = Mock()
+        del client.localNode
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_client_no_send_data_returns_early(self, mock_settings):
+        mock_settings.return_value = (True, 30.0)
+        client = Mock(spec=["localNode"])
+        client.localNode = Mock()
+        del client.sendData
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_send_data_not_callable_returns_early(self, mock_settings):
+        mock_settings.return_value = (True, 30.0)
+        client = Mock()
+        client.localNode = Mock()
+        client.sendData = "not_callable"
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+    @patch("mmrelay.meshtastic_utils._submit_metadata_probe")
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_degraded_executor_returns_early(self, mock_settings, mock_submit):
+        mock_settings.return_value = (True, 30.0)
+        mock_submit.side_effect = mu.MetadataExecutorDegradedError("degraded")
+        client = Mock()
+        client.localNode = Mock()
+        client.sendData = Mock()
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+    @patch("mmrelay.meshtastic_utils._submit_metadata_probe")
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_runtime_error_returns_early(self, mock_settings, mock_submit):
+        mock_settings.return_value = (True, 30.0)
+        mock_submit.side_effect = RuntimeError("executor broken")
+        client = Mock()
+        client.localNode = Mock()
+        client.sendData = Mock()
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+    @patch("mmrelay.meshtastic_utils._submit_metadata_probe")
+    @patch("mmrelay.meshtastic_utils._get_connect_time_probe_settings")
+    def test_probe_future_none_returns_early(self, mock_settings, mock_submit):
+        mock_settings.return_value = (True, 30.0)
+        mock_submit.return_value = None
+        client = Mock()
+        client.localNode = Mock()
+        client.sendData = Mock()
+
+        mu._schedule_connect_time_calibration_probe(
+            client,
+            connection_type=CONNECTION_TYPE_TCP,
+            active_config={"meshtastic": {}},
+        )
+
+
+# ---------------------------------------------------------------------------
+# Absorbed from tests/test_meshtastic_utils_node_name_refresh.py
+# ---------------------------------------------------------------------------
+
+
+class _OnePassEvent:
+    """Event that starts cleared and sets itself when awaited once."""
+
+    def __init__(self) -> None:
+        self._set = False
+
+    def is_set(self) -> bool:
+        return self._set
+
+    async def wait(self) -> None:
+        self._set = True
+
+
+class _TimeoutThenSetEvent:
+    """Event whose first wait times out and second wait sets the event."""
+
+    def __init__(self) -> None:
+        self._set = False
+        self._wait_calls = 0
+        self.first_wait_cancelled = False
+
+    def is_set(self) -> bool:
+        return self._set
+
+    async def wait(self) -> None:
+        self._wait_calls += 1
+        if self._wait_calls == 1:
+            try:
+                # Sleep long enough to be cancelled by the refresh loop's timeout
+                await asyncio.sleep(1)
+            except asyncio.CancelledError:
+                self.first_wait_cancelled = True
+                raise
+            return
+        self._set = True
+
+
+class _ClientWithoutNodes:
+    """Minimal client shape with no nodes attribute."""
+
+
+class _ClientWithNodes:
+    """Minimal client shape with a dict-backed nodes attribute."""
+
+    def __init__(self, nodes: dict[str, Any]) -> None:
+        self.nodes = nodes
+
+
+def test_get_nodedb_refresh_interval_ignores_non_dict_config() -> None:
+    """Non-dict config inputs should fall back to default nodedb refresh interval."""
+    interval = mu.get_nodedb_refresh_interval_seconds(cast(Any, []))
+    assert interval == DEFAULT_NODEDB_REFRESH_INTERVAL
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_skips_when_nodes_attribute_unavailable() -> (
+    None
+):
+    """Missing client.nodes should skip sync rather than treat as empty nodedb."""
+    with (
+        patch.object(mu, "meshtastic_client", _ClientWithoutNodes()),
+        patch.object(mu, "sync_name_tables_if_changed") as mock_sync,
+    ):
+        await mu.refresh_node_name_tables(
+            _OnePassEvent(),  # type: ignore[arg-type]
+            refresh_interval_seconds=0.01,
+        )
+    mock_sync.assert_not_called()
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_handles_timeout_then_retries() -> None:
+    """Refresh loop should continue after wait timeout and retry sync."""
+    event = _TimeoutThenSetEvent()
+    client = _ClientWithNodes(
+        {
+            "node_a": {
+                "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
+            }
+        }
+    )
+    state_after_first_refresh = {"state": 1}
+    state_after_second_refresh = {"state": 2}
+    with (
+        patch.object(mu, "meshtastic_client", client),
+        patch.object(
+            mu,
+            "sync_name_tables_if_changed",
+            side_effect=[state_after_first_refresh, state_after_second_refresh],
+        ) as mock_sync,
+    ):
+        await mu.refresh_node_name_tables(
+            event,  # type: ignore[arg-type]
+            refresh_interval_seconds=0.01,
+        )
+    assert event.first_wait_cancelled is True
+    assert event._wait_calls == 2
+    assert mock_sync.call_count == 2
+    assert mock_sync.call_args_list[0].args[1] is None
+    assert mock_sync.call_args_list[1].args[1] is state_after_first_refresh
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_logs_unavailable_only_once_per_state() -> None:
+    """Unavailable-client debug noise should be deduplicated across loop iterations."""
+    event = _TimeoutThenSetEvent()
+    with (
+        patch.object(mu, "meshtastic_client", None),
+        patch.object(mu, "logger") as mock_logger,
+    ):
+        await mu.refresh_node_name_tables(
+            event,  # type: ignore[arg-type]
+            refresh_interval_seconds=0.01,
+        )
+
+    unavailable_calls = [
+        c
+        for c in mock_logger.debug.call_args_list
+        if c.args
+        and c.args[0]
+        == "Skipping name-cache refresh from NodeDB because Meshtastic client is unavailable"
+    ]
+    assert len(unavailable_calls) == 1
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_uses_reconnecting_unavailable_message() -> None:
+    """When reconnecting, refresh logs reconnect-specific unavailability reason."""
+    event = _OnePassEvent()
+    with (
+        patch.object(mu, "meshtastic_client", None),
+        patch.object(mu, "reconnecting", True),
+        patch.object(mu, "logger") as mock_logger,
+    ):
+        await mu.refresh_node_name_tables(
+            event,  # type: ignore[arg-type]
+            refresh_interval_seconds=0.01,
+        )
+
+    mock_logger.debug.assert_any_call(
+        "Skipping name-cache refresh from NodeDB while reconnection is in progress"
+    )
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_non_positive_interval_exits_after_one_pass() -> (
+    None
+):
+    """Zero interval should perform one immediate pass and return."""
+    with (
+        patch.object(
+            mu,
+            "meshtastic_client",
+            _ClientWithNodes(
+                {
+                    "node_a": {
+                        "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
+                    }
+                }
+            ),
+        ),
+        patch.object(mu, "sync_name_tables_if_changed", return_value=()) as mock_sync,
+    ):
+        await mu.refresh_node_name_tables(
+            _OnePassEvent(),  # type: ignore[arg-type]
+            refresh_interval_seconds=0.0,
+        )
+    mock_sync.assert_called_once()
+
+
+@pytest.mark.asyncio
+async def test_refresh_node_name_tables_handles_sync_exceptions(caplog) -> None:
+    """Sync errors should be logged and re-raised for supervisor handling."""
+    caplog.set_level(logging.ERROR, logger=mu.logger.name)
+    original_propagate = mu.logger.propagate
+    mu.logger.propagate = True
+    client = _ClientWithNodes(
+        {
+            "node_a": {
+                "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
+            }
+        }
+    )
+    try:
+        with (
+            patch.object(mu, "meshtastic_client", client),
+            patch.object(
+                mu,
+                "sync_name_tables_if_changed",
+                side_effect=RuntimeError("sync failure"),
+            ) as mock_sync,
+        ):
+            with pytest.raises(RuntimeError, match="sync failure"):
+                await mu.refresh_node_name_tables(
+                    _OnePassEvent(),  # type: ignore[arg-type]
+                    refresh_interval_seconds=0.0,
+                )
+    finally:
+        mu.logger.propagate = original_propagate
+    mock_sync.assert_called_once()
+    assert any(
+        "Failed to refresh name-cache tables from NodeDB snapshot" in record.message
+        for record in caplog.records
+    )
