--- tests/test_meshtastic_utils_node_name_refresh.py
@@ -1,242 +1,0 @@
-#!/usr/bin/env python3
-"""Targeted tests for node-name refresh interval and refresh-loop edge paths."""
-
-import asyncio
-import logging
-from typing import Any, cast
-from unittest.mock import patch
-
-import pytest
-
-import mmrelay.meshtastic_utils as mu
-from mmrelay.constants.config import DEFAULT_NODEDB_REFRESH_INTERVAL
-
-
-class _OnePassEvent:
-    """Event that starts cleared and sets itself when awaited once."""
-
-    def __init__(self) -> None:
-        self._set = False
-
-    def is_set(self) -> bool:
-        return self._set
-
-    async def wait(self) -> None:
-        self._set = True
-
-
-class _TimeoutThenSetEvent:
-    """Event whose first wait times out and second wait sets the event."""
-
-    def __init__(self) -> None:
-        self._set = False
-        self._wait_calls = 0
-        self.first_wait_cancelled = False
-
-    def is_set(self) -> bool:
-        return self._set
-
-    async def wait(self) -> None:
-        self._wait_calls += 1
-        if self._wait_calls == 1:
-            try:
-                # Sleep long enough to be cancelled by the refresh loop's timeout
-                await asyncio.sleep(1)
-            except asyncio.CancelledError:
-                self.first_wait_cancelled = True
-                raise
-            return
-        self._set = True
-
-
-class _ClientWithoutNodes:
-    """Minimal client shape with no nodes attribute."""
-
-
-class _ClientWithNodes:
-    """Minimal client shape with a dict-backed nodes attribute."""
-
-    def __init__(self, nodes: dict[str, Any]) -> None:
-        self.nodes = nodes
-
-
-def test_get_nodedb_refresh_interval_ignores_non_dict_config(
-    reset_meshtastic_globals,
-) -> None:
-    """Non-dict config inputs should fall back to default nodedb refresh interval."""
-    _ = reset_meshtastic_globals
-    interval = mu.get_nodedb_refresh_interval_seconds(cast(Any, []))
-    assert interval == DEFAULT_NODEDB_REFRESH_INTERVAL
-
-
-def test_refresh_node_name_tables_skips_when_nodes_attribute_unavailable(
-    reset_meshtastic_globals,
-) -> None:
-    """Missing client.nodes should skip sync rather than treat as empty nodedb."""
-    _ = reset_meshtastic_globals
-    with (
-        patch.object(mu, "meshtastic_client", _ClientWithoutNodes()),
-        patch.object(mu, "sync_name_tables_if_changed") as mock_sync,
-    ):
-        asyncio.run(
-            mu.refresh_node_name_tables(
-                _OnePassEvent(),  # type: ignore[arg-type]
-                refresh_interval_seconds=0.01,
-            )
-        )
-    mock_sync.assert_not_called()
-
-
-def test_refresh_node_name_tables_handles_timeout_then_retries(
-    reset_meshtastic_globals,
-) -> None:
-    """Refresh loop should continue after wait timeout and retry sync."""
-    _ = reset_meshtastic_globals
-    event = _TimeoutThenSetEvent()
-    client = _ClientWithNodes(
-        {
-            "node_a": {
-                "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
-            }
-        }
-    )
-    state_after_first_refresh = {"state": 1}
-    state_after_second_refresh = {"state": 2}
-    with (
-        patch.object(mu, "meshtastic_client", client),
-        patch.object(
-            mu,
-            "sync_name_tables_if_changed",
-            side_effect=[state_after_first_refresh, state_after_second_refresh],
-        ) as mock_sync,
-    ):
-        asyncio.run(
-            mu.refresh_node_name_tables(
-                event,  # type: ignore[arg-type]
-                refresh_interval_seconds=0.01,
-            )
-        )
-    assert event.first_wait_cancelled is True
-    assert event._wait_calls == 2
-    assert mock_sync.call_count == 2
-    assert mock_sync.call_args_list[0].args[1] is None
-    assert mock_sync.call_args_list[1].args[1] is state_after_first_refresh
-
-
-def test_refresh_node_name_tables_logs_unavailable_only_once_per_state(
-    reset_meshtastic_globals,
-) -> None:
-    """Unavailable-client debug noise should be deduplicated across loop iterations."""
-    _ = reset_meshtastic_globals
-    event = _TimeoutThenSetEvent()
-    with (
-        patch.object(mu, "meshtastic_client", None),
-        patch.object(mu, "logger") as mock_logger,
-    ):
-        asyncio.run(
-            mu.refresh_node_name_tables(
-                event,  # type: ignore[arg-type]
-                refresh_interval_seconds=0.01,
-            )
-        )
-
-    unavailable_calls = [
-        c
-        for c in mock_logger.debug.call_args_list
-        if c.args
-        and c.args[0]
-        == "Skipping name-cache refresh from NodeDB because Meshtastic client is unavailable"
-    ]
-    assert len(unavailable_calls) == 1
-
-
-def test_refresh_node_name_tables_uses_reconnecting_unavailable_message(
-    reset_meshtastic_globals,
-) -> None:
-    """When reconnecting, refresh logs reconnect-specific unavailability reason."""
-    _ = reset_meshtastic_globals
-    event = _OnePassEvent()
-    with (
-        patch.object(mu, "meshtastic_client", None),
-        patch.object(mu, "reconnecting", True),
-        patch.object(mu, "logger") as mock_logger,
-    ):
-        asyncio.run(
-            mu.refresh_node_name_tables(
-                event,  # type: ignore[arg-type]
-                refresh_interval_seconds=0.01,
-            )
-        )
-
-    mock_logger.debug.assert_any_call(
-        "Skipping name-cache refresh from NodeDB while reconnection is in progress"
-    )
-
-
-def test_refresh_node_name_tables_non_positive_interval_exits_after_one_pass(
-    reset_meshtastic_globals,
-) -> None:
-    """Zero interval should perform one immediate pass and return."""
-    _ = reset_meshtastic_globals
-    with (
-        patch.object(
-            mu,
-            "meshtastic_client",
-            _ClientWithNodes(
-                {
-                    "node_a": {
-                        "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
-                    }
-                }
-            ),
-        ),
-        patch.object(mu, "sync_name_tables_if_changed", return_value=()) as mock_sync,
-    ):
-        asyncio.run(
-            mu.refresh_node_name_tables(
-                _OnePassEvent(),  # type: ignore[arg-type]
-                refresh_interval_seconds=0.0,
-            )
-        )
-    mock_sync.assert_called_once()
-
-
-def test_refresh_node_name_tables_handles_sync_exceptions(
-    reset_meshtastic_globals,
-    caplog,
-) -> None:
-    """Sync errors should be logged and re-raised for supervisor handling."""
-    _ = reset_meshtastic_globals
-    caplog.set_level(logging.ERROR, logger=mu.logger.name)
-    original_propagate = mu.logger.propagate
-    mu.logger.propagate = True
-    client = _ClientWithNodes(
-        {
-            "node_a": {
-                "user": {"id": "!1", "longName": "Alpha", "shortName": "A"},
-            }
-        }
-    )
-    try:
-        with (
-            patch.object(mu, "meshtastic_client", client),
-            patch.object(
-                mu,
-                "sync_name_tables_if_changed",
-                side_effect=RuntimeError("sync failure"),
-            ) as mock_sync,
-        ):
-            with pytest.raises(RuntimeError, match="sync failure"):
-                asyncio.run(
-                    mu.refresh_node_name_tables(
-                        _OnePassEvent(),  # type: ignore[arg-type]
-                        refresh_interval_seconds=0.0,
-                    )
-                )
-    finally:
-        mu.logger.propagate = original_propagate
-    mock_sync.assert_called_once()
-    assert any(
-        "Failed to refresh name-cache tables from NodeDB snapshot" in record.message
-        for record in caplog.records
-    )
